+++ src/configuration.ts
@@ -19,7 +19,6 @@ import * as contextAddresses from "@mangrovedao/context-addresses";
 import * as eth from "./eth";
 import clone from "just-clone";
 import deepmerge from "deepmerge";
-import moize from "moize";
 import semver from "semver";
 
 // Make keys optional at all levels of T
@@ -314,16 +313,6 @@ export const tokensConfiguration = {
     tokensConfiguration.setDecimals(tokenName, decimals);
     return decimals;
   },
-
-  /**
-   * Read chain for decimals of `address` on current network
-   */
-  fetchDecimalsFromAddress: moize(
-    async (address: string, provider: Provider): Promise<number> => {
-      const token = typechain.IERC20__factory.connect(address, provider);
-      return token.decimals();
-    },
-  ),
 
   /**
    * Read displayed decimals for `tokenName`.